# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "JffEhr.Api",
    "version": "1.0"
  },
  "paths": {
    "/api/AuditLogs": {
      "get": {
        "tags": [
          "AuditLogs"
        ],
        "parameters": [
          {
            "name": "entityTable",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "entityId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "userId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditLogDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditLogDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditLogDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/AuditLogs/{id}": {
      "get": {
        "tags": [
          "AuditLogs"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AuditLogDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditLogDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditLogDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/CampRegistrations": {
      "get": {
        "tags": [
          "CampRegistrations"
        ],
        "parameters": [
          {
            "name": "campId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "camperId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CampRegistrationDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CampRegistrationDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CampRegistrationDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "CampRegistrations"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCampRegistrationRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCampRegistrationRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCampRegistrationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CampRegistrationDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampRegistrationDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampRegistrationDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/CampRegistrations/{id}": {
      "get": {
        "tags": [
          "CampRegistrations"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CampRegistrationDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampRegistrationDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampRegistrationDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "CampRegistrations"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCampRegistrationRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCampRegistrationRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCampRegistrationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "CampRegistrations"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/Camps/{id}": {
      "get": {
        "tags": [
          "Camps"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CampDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Camps"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCampRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCampRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCampRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Camps"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AuditLogDto": {
        "type": "object",
        "properties": {
          "auditId": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "userEmail": {
            "type": "string",
            "nullable": true
          },
          "entityTable": {
            "type": "string",
            "nullable": true
          },
          "entityId": {
            "type": "string",
            "format": "uuid"
          },
          "action": {
            "type": "string",
            "nullable": true
          },
          "beforeState": {
            "type": "string",
            "nullable": true
          },
          "afterState": {
            "type": "string",
            "nullable": true
          },
          "ipAddress": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "CampDto": {
        "required": [
          "campType",
          "province",
          "status",
          "venue"
        ],
        "type": "object",
        "properties": {
          "campId": {
            "type": "string",
            "format": "uuid"
          },
          "campNumber": {
            "type": "integer",
            "format": "int32"
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "format": "date"
          },
          "venue": {
            "type": "string",
            "nullable": true
          },
          "province": {
            "type": "string",
            "nullable": true
          },
          "campType": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "CampRegistrationDto": {
        "required": [
          "status"
        ],
        "type": "object",
        "properties": {
          "registrationId": {
            "type": "string",
            "format": "uuid"
          },
          "campId": {
            "type": "string",
            "format": "uuid"
          },
          "camperId": {
            "type": "string",
            "format": "uuid"
          },
          "cabin": {
            "type": "string",
            "nullable": true
          },
          "groupName": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "registeredAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "CreateCampRegistrationRequest": {
        "required": [
          "camperId",
          "campId",
          "status"
        ],
        "type": "object",
        "properties": {
          "campId": {
            "type": "string",
            "format": "uuid"
          },
          "camperId": {
            "type": "string",
            "format": "uuid"
          },
          "cabin": {
            "maxLength": 20,
            "type": "string",
            "nullable": true
          },
          "groupName": {
            "maxLength": 40,
            "type": "string",
            "nullable": true
          },
          "status": {
            "maxLength": 20,
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "UpdateCampRegistrationRequest": {
        "required": [
          "status"
        ],
        "type": "object",
        "properties": {
          "cabin": {
            "maxLength": 20,
            "type": "string",
            "nullable": true
          },
          "groupName": {
            "maxLength": 40,
            "type": "string",
            "nullable": true
          },
          "status": {
            "maxLength": 20,
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "UpdateCampRequest": {
        "required": [
          "campType",
          "endDate",
          "province",
          "startDate",
          "status",
          "venue"
        ],
        "type": "object",
        "properties": {
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "format": "date"
          },
          "venue": {
            "maxLength": 200,
            "minLength": 1,
            "type": "string"
          },
          "province": {
            "maxLength": 40,
            "minLength": 1,
            "type": "string"
          },
          "campType": {
            "maxLength": 60,
            "minLength": 1,
            "type": "string"
          },
          "status": {
            "maxLength": 20,
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}